backend/requirements.txt
django==4.2.30
djangorestframework==3.16.1
django-cors-headers==4.9.0
whitenoise==6.11.0
gunicorn==23.0.0
